Multi-Agent-Custom-Automation-Engine-Solution-Accelerator/src/mcp_server/requirements.txt
fastmcp==2.13.0
uvicorn[standard]==0.38.0
pydantic==2.11.7
pydantic-settings==2.6.1
httpx==0.28.1
opentelemetry-distro==0.49b0
opentelemetry-exporter-otlp==1.28.0
opentelemetry-instrumentation-httpx==0.49b0